Implement swank:frame-locals-and-catch-tags
This allows the display of local variables of a frame if there are
any.  We now store the call chain in a variable *recent-call-chain*
right when we catch the exception, so hopefully it won't get garbled
by our later evaluations of the swank module itself.

Also corrected the frame reversal, we missed the topmost frame which
holds the info we wanted to know.

The frameinfo parsing has been taken from csi.scm from chicken-core.